feat(i2c): Add I2cMenu to i2c component * Move I2cMenu from component example into component so that others can use it more effectively * Update menu to allow easier reuse / subclassing and to be within the espp namespace * Add protection around menu so that it allows compilation if exceptions are disabled.
