+++ components/i2c/CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
   INCLUDE_DIRS "include"
-  REQUIRES "driver" "base_component"
+  REQUIRES "driver" "base_component" "cli"
   )
